Fix asymmetric error behaviour
The erlang zookeeper library, ezk, tries to handle errors by monitoring
the erlang process that uses the connection. This is to ensure that the
connection is shutdown if the using process for some reason terminates.
However, if the connection dies by itself, there is no mechanism to inform
the using process that there no longer exists a connection and it will
experience a crash the next time it is executing any ezk commands, since
it uses gen_server:call/3.